requirements.txt
flask
transformers
torch
faiss-cpu
scikit-learn
numpy
Flask-Cors